Split routes up into different concerns
